+++ src/abi/index.test.ts
@@ -1,6 +1,6 @@
 import { expect, test } from 'vitest'
 
-import * as abi from './abi.js'
+import * as abi from './index.js'
 
 test('exports abi utils', () => {
   expect(Object.keys(abi)).toMatchInlineSnapshot(`
