+++ src/chains/package.json
@@ -1,0 +1,6 @@
+{
+  "type": "module",
+  "types": "../_types/chains/index.d.ts",
+  "module": "../_esm/chains/index.js",
+  "main": "../_cjs/chains/index.js"
+}
