+++ src/chains/utils/package.json
@@ -1,0 +1,6 @@
+{
+  "type": "module",
+  "types": "../../_types/chains/utils/index.d.ts",
+  "module": "../../_esm/chains/utils/index.js",
+  "main": "../../_cjs/chains/utils/index.js"
+}
