+++ src/contract/index.ts
@@ -2,27 +2,27 @@ export {
   type CreateContractEventFilterParameters,
   type CreateContractEventFilterReturnType,
   createContractEventFilter,
-} from './actions/public/createContractEventFilter.js'
+} from '../actions/public/createContractEventFilter.js'
 export {
   type EstimateContractGasParameters,
   type EstimateContractGasReturnType,
   estimateContractGas,
-} from './actions/public/estimateContractGas.js'
+} from '../actions/public/estimateContractGas.js'
 export {
   type GetBytecodeParameters,
   type GetBytecodeReturnType,
   getBytecode,
-} from './actions/public/getBytecode.js'
+} from '../actions/public/getBytecode.js'
 export {
   type GetStorageAtParameters,
   type GetStorageAtReturnType,
   getStorageAt,
-} from './actions/public/getStorageAt.js'
+} from '../actions/public/getStorageAt.js'
 export {
   type MulticallParameters,
   type MulticallReturnType,
   multicall,
-} from './actions/public/multicall.js'
+} from '../actions/public/multicall.js'
 export type {
   WatchEventOnLogsFn,
   /** @deprecated - use `WatchEventOnLogsFn` instead. */
@@ -30,81 +30,81 @@ export type {
   WatchEventOnLogsParameter,
   /** @deprecated - use `WatchEventOnLogsParameter` instead. */
   WatchEventOnLogsParameter as OnLogsParameter,
-} from './actions/public/watchEvent.js'
+} from '../actions/public/watchEvent.js'
 export {
   type ReadContractParameters,
   type ReadContractReturnType,
   readContract,
-} from './actions/public/readContract.js'
+} from '../actions/public/readContract.js'
 export {
   type SimulateContractParameters,
   type SimulateContractReturnType,
   simulateContract,
-} from './actions/public/simulateContract.js'
+} from '../actions/public/simulateContract.js'
 export {
   type WatchContractEventParameters,
   watchContractEvent,
-} from './actions/public/watchContractEvent.js'
+} from '../actions/public/watchContractEvent.js'
 export {
   type DeployContractParameters,
   type DeployContractReturnType,
   deployContract,
-} from './actions/wallet/deployContract.js'
+} from '../actions/wallet/deployContract.js'
 export {
   type WriteContractParameters,
   type WriteContractReturnType,
   writeContract,
-} from './actions/wallet/writeContract.js'
+} from '../actions/wallet/writeContract.js'
 export {
   type DecodeAbiParametersReturnType,
   decodeAbiParameters,
-} from './utils/abi/decodeAbiParameters.js'
+} from '../utils/abi/decodeAbiParameters.js'
 export {
   type DecodeErrorResultParameters,
   type DecodeErrorResultReturnType,
   decodeErrorResult,
-} from './utils/abi/decodeErrorResult.js'
+} from '../utils/abi/decodeErrorResult.js'
 export {
   type DecodeEventLogParameters,
   type DecodeEventLogReturnType,
   decodeEventLog,
-} from './utils/abi/decodeEventLog.js'
+} from '../utils/abi/decodeEventLog.js'
 export {
   type DecodeFunctionDataParameters,
   decodeFunctionData,
-} from './utils/abi/decodeFunctionData.js'
+} from '../utils/abi/decodeFunctionData.js'
 export {
   type DecodeFunctionResultParameters,
   type DecodeFunctionResultReturnType,
   decodeFunctionResult,
-} from './utils/abi/decodeFunctionResult.js'
+} from '../utils/abi/decodeFunctionResult.js'
 export {
   type EncodeAbiParametersReturnType,
   encodeAbiParameters,
-} from './utils/abi/encodeAbiParameters.js'
+} from '../utils/abi/encodeAbiParameters.js'
 export {
   type EncodeDeployDataParameters,
   encodeDeployData,
-} from './utils/abi/encodeDeployData.js'
+} from '../utils/abi/encodeDeployData.js'
 export {
   type EncodeErrorResultParameters,
   encodeErrorResult,
-} from './utils/abi/encodeErrorResult.js'
+} from '../utils/abi/encodeErrorResult.js'
 export {
   type EncodeEventTopicsParameters,
   encodeEventTopics,
-} from './utils/abi/encodeEventTopics.js'
+} from '../utils/abi/encodeEventTopics.js'
 export {
   type EncodeFunctionDataParameters,
   encodeFunctionData,
-} from './utils/abi/encodeFunctionData.js'
+} from '../utils/abi/encodeFunctionData.js'
 export {
   type EncodeFunctionResultParameters,
   encodeFunctionResult,
-} from './utils/abi/encodeFunctionResult.js'
+} from '../utils/abi/encodeFunctionResult.js'
 export {
   type GetAbiItemParameters,
   getAbiItem,
-} from './utils/abi/getAbiItem.js'
-export { formatAbiItemWithArgs } from './utils/abi/formatAbiItemWithArgs.js'
-export { formatAbiItem } from './utils/abi/formatAbiItem.js'
+} from '../utils/abi/getAbiItem.js'
+export { formatAbiItemWithArgs } from '../utils/abi/formatAbiItemWithArgs.js'
+export { formatAbiItem } from '../utils/abi/formatAbiItem.js'
