+++ src/ens/index.ts
@@ -1,28 +1,28 @@
-export { normalize } from './utils/ens/normalize.js'
+export { normalize } from '../utils/ens/normalize.js'
 export {
   type GetEnsAddressParameters,
   type GetEnsAddressReturnType,
   getEnsAddress,
-} from './actions/ens/getEnsAddress.js'
+} from '../actions/ens/getEnsAddress.js'
 export {
   type GetEnsAvatarParameters,
   type GetEnsAvatarReturnType,
   getEnsAvatar,
-} from './actions/ens/getEnsAvatar.js'
+} from '../actions/ens/getEnsAvatar.js'
 export {
   type GetEnsNameParameters,
   type GetEnsNameReturnType,
   getEnsName,
-} from './actions/ens/getEnsName.js'
+} from '../actions/ens/getEnsName.js'
 export {
   type GetEnsResolverParameters,
   type GetEnsResolverReturnType,
   getEnsResolver,
-} from './actions/ens/getEnsResolver.js'
+} from '../actions/ens/getEnsResolver.js'
 export {
   type GetEnsTextParameters,
   type GetEnsTextReturnType,
   getEnsText,
-} from './actions/ens/getEnsText.js'
-export { labelhash } from './utils/ens/labelhash.js'
-export { namehash } from './utils/ens/namehash.js'
+} from '../actions/ens/getEnsText.js'
+export { labelhash } from '../utils/ens/labelhash.js'
+export { namehash } from '../utils/ens/namehash.js'
