+++ src/ens/package.json
@@ -1,0 +1,6 @@
+{
+  "type": "module",
+  "types": "../_types/ens/index.d.ts",
+  "module": "../_esm/ens/index.js",
+  "main": "../_cjs/ens/index.js"
+}
