+++ src/package.json
@@ -23,9 +23,9 @@
       "default": "./_cjs/index.js"
     },
     "./abi": {
-      "types": "./_types/abi.d.ts",
-      "import": "./_esm/abi.js",
-      "default": "./_cjs/abi.js"
+      "types": "./_types/abi/index.d.ts",
+      "import": "./_esm/abi/index.js",
+      "default": "./_cjs/abi/index.js"
     },
     "./accounts": {
       "types": "./_types/accounts/index.d.ts",
@@ -43,29 +43,29 @@
       "default": "./_cjs/chains/index.js"
     },
     "./chains/utils": {
-      "types": "./_types/chains/utils.d.ts",
-      "import": "./_esm/chains/utils.js",
-      "default": "./_cjs/chains/utils.js"
+      "types": "./_types/chains/utils/index.d.ts",
+      "import": "./_esm/chains/utils/index.js",
+      "default": "./_cjs/chains/utils/index.js"
     },
     "./contract": {
-      "types": "./_types/contract.d.ts",
-      "import": "./_esm/contract.js",
-      "default": "./_cjs/contract.js"
+      "types": "./_types/contract/index.d.ts",
+      "import": "./_esm/contract/index.js",
+      "default": "./_cjs/contract/index.js"
     },
     "./ens": {
-      "types": "./_types/ens.d.ts",
-      "import": "./_esm/ens.js",
-      "default": "./_cjs/ens.js"
+      "types": "./_types/ens/index.d.ts",
+      "import": "./_esm/ens/index.js",
+      "default": "./_cjs/ens/index.js"
     },
     "./public": {
-      "types": "./_types/public.d.ts",
-      "import": "./_esm/public.js",
-      "default": "./_cjs/public.js"
+      "types": "./_types/public/index.d.ts",
+      "import": "./_esm/public/index.js",
+      "default": "./_cjs/public/index.js"
     },
     "./test": {
-      "types": "./_types/test.d.ts",
-      "import": "./_esm/test.js",
-      "default": "./_cjs/test.js"
+      "types": "./_types/test/index.d.ts",
+      "import": "./_esm/test/index.js",
+      "default": "./_cjs/test/index.js"
     },
     "./utils": {
       "types": "./_types/utils/index.d.ts",
@@ -73,14 +73,14 @@
       "default": "./_cjs/utils/index.js"
     },
     "./wallet": {
-      "types": "./_types/wallet.d.ts",
-      "import": "./_esm/wallet.js",
-      "default": "./_cjs/wallet.js"
+      "types": "./_types/wallet/index.d.ts",
+      "import": "./_esm/wallet/index.js",
+      "default": "./_cjs/wallet/index.js"
     },
     "./window": {
-      "types": "./_types/window.d.ts",
-      "import": "./_esm/window.js",
-      "default": "./_cjs/window.js"
+      "types": "./_types/window/index.d.ts",
+      "import": "./_esm/window/index.js",
+      "default": "./_cjs/window/index.js"
     },
     "./package.json": "./package.json"
   },
