+++ src/public/index.ts
@@ -2,126 +2,126 @@ export {
   call,
   type CallParameters,
   type CallReturnType,
-} from './actions/public/call.js'
+} from '../actions/public/call.js'
 export {
   createBlockFilter,
   type CreateBlockFilterReturnType,
-} from './actions/public/createBlockFilter.js'
+} from '../actions/public/createBlockFilter.js'
 export {
   createEventFilter,
   type CreateEventFilterParameters,
   type CreateEventFilterReturnType,
-} from './actions/public/createEventFilter.js'
+} from '../actions/public/createEventFilter.js'
 export {
   createPendingTransactionFilter,
   type CreatePendingTransactionFilterReturnType,
-} from './actions/public/createPendingTransactionFilter.js'
+} from '../actions/public/createPendingTransactionFilter.js'
 export {
   estimateGas,
   type EstimateGasParameters,
   type EstimateGasReturnType,
-} from './actions/public/estimateGas.js'
+} from '../actions/public/estimateGas.js'
 export {
   getBalance,
   type GetBalanceParameters,
   type GetBalanceReturnType,
-} from './actions/public/getBalance.js'
+} from '../actions/public/getBalance.js'
 export {
   getBlock,
   type GetBlockParameters,
   type GetBlockReturnType,
-} from './actions/public/getBlock.js'
+} from '../actions/public/getBlock.js'
 export {
   getBlockNumber,
   getBlockNumberCache,
   type GetBlockNumberParameters,
   type GetBlockNumberReturnType,
-} from './actions/public/getBlockNumber.js'
+} from '../actions/public/getBlockNumber.js'
 export {
   getBlockTransactionCount,
   type GetBlockTransactionCountParameters,
   type GetBlockTransactionCountReturnType,
-} from './actions/public/getBlockTransactionCount.js'
+} from '../actions/public/getBlockTransactionCount.js'
 export {
   getBytecode,
   type GetBytecodeParameters,
   type GetBytecodeReturnType,
-} from './actions/public/getBytecode.js'
-export { getChainId } from './actions/public/getChainId.js'
+} from '../actions/public/getBytecode.js'
+export { getChainId } from '../actions/public/getChainId.js'
 export {
   getFeeHistory,
   type GetFeeHistoryParameters,
   type GetFeeHistoryReturnType,
-} from './actions/public/getFeeHistory.js'
+} from '../actions/public/getFeeHistory.js'
 export {
   estimateFeesPerGas,
   type EstimateFeesPerGasParameters,
   type EstimateFeesPerGasReturnType,
-} from './actions/public/estimateFeesPerGas.js'
+} from '../actions/public/estimateFeesPerGas.js'
 export {
   getFilterChanges,
   type GetFilterChangesParameters,
   type GetFilterChangesReturnType,
-} from './actions/public/getFilterChanges.js'
+} from '../actions/public/getFilterChanges.js'
 export {
   getFilterLogs,
   type GetFilterLogsParameters,
   type GetFilterLogsReturnType,
-} from './actions/public/getFilterLogs.js'
-export { getLogs } from './actions/public/getLogs.js'
+} from '../actions/public/getFilterLogs.js'
+export { getLogs } from '../actions/public/getLogs.js'
 export {
   getGasPrice,
   type GetGasPriceReturnType,
-} from './actions/public/getGasPrice.js'
+} from '../actions/public/getGasPrice.js'
 export {
   estimateMaxPriorityFeePerGas,
   type EstimateMaxPriorityFeePerGasParameters,
   type EstimateMaxPriorityFeePerGasReturnType,
-} from './actions/public/estimateMaxPriorityFeePerGas.js'
+} from '../actions/public/estimateMaxPriorityFeePerGas.js'
 export {
   getTransaction,
   type GetTransactionParameters,
   type GetTransactionReturnType,
-} from './actions/public/getTransaction.js'
+} from '../actions/public/getTransaction.js'
 export {
   getTransactionConfirmations,
   type GetTransactionConfirmationsParameters,
   type GetTransactionConfirmationsReturnType,
-} from './actions/public/getTransactionConfirmations.js'
+} from '../actions/public/getTransactionConfirmations.js'
 export {
   getTransactionCount,
   type GetTransactionCountParameters,
   type GetTransactionCountReturnType,
-} from './actions/public/getTransactionCount.js'
+} from '../actions/public/getTransactionCount.js'
 export {
   getTransactionReceipt,
   type GetTransactionReceiptParameters,
   type GetTransactionReceiptReturnType,
-} from './actions/public/getTransactionReceipt.js'
+} from '../actions/public/getTransactionReceipt.js'
 export {
   uninstallFilter,
   type UninstallFilterParameters,
   type UninstallFilterReturnType,
-} from './actions/public/uninstallFilter.js'
+} from '../actions/public/uninstallFilter.js'
 export {
   waitForTransactionReceipt,
   type ReplacementReason,
   type ReplacementReturnType,
   type WaitForTransactionReceiptParameters,
   type WaitForTransactionReceiptReturnType,
-} from './actions/public/waitForTransactionReceipt.js'
+} from '../actions/public/waitForTransactionReceipt.js'
 export {
   watchBlockNumber,
   type OnBlockNumberFn,
   type OnBlockNumberParameter,
   type WatchBlockNumberParameters,
-} from './actions/public/watchBlockNumber.js'
+} from '../actions/public/watchBlockNumber.js'
 export {
   watchBlocks,
   type OnBlock,
   type OnBlockParameter,
   type WatchBlocksParameters,
-} from './actions/public/watchBlocks.js'
+} from '../actions/public/watchBlocks.js'
 export {
   watchEvent,
   type WatchEventOnLogsFn,
@@ -130,10 +130,10 @@ export {
   type WatchEventOnLogsParameter,
   /** @deprecated - use `WatchEventOnLogsParameter` instead. */
   type WatchEventOnLogsParameter as OnLogsParameter,
-} from './actions/public/watchEvent.js'
+} from '../actions/public/watchEvent.js'
 export {
   watchPendingTransactions,
   type OnTransactionsFn,
   type OnTransactionsParameter,
   type WatchPendingTransactionsParameters,
-} from './actions/public/watchPendingTransactions.js'
+} from '../actions/public/watchPendingTransactions.js'
