+++ src/public/package.json
@@ -1,0 +1,6 @@
+{
+  "type": "module",
+  "types": "../_types/public/index.d.ts",
+  "module": "../_esm/public/index.js",
+  "main": "../_cjs/public/index.js"
+}
