--- src/test.ts
@@ -1,83 +1,0 @@
-export {
-  dropTransaction,
-  type DropTransactionParameters,
-} from './actions/test/dropTransaction.js'
-export {
-  getAutomine,
-  type GetAutomineReturnType,
-} from './actions/test/getAutomine.js'
-export {
-  getTxpoolContent,
-  type GetTxpoolContentReturnType,
-} from './actions/test/getTxpoolContent.js'
-export {
-  getTxpoolStatus,
-  type GetTxpoolStatusReturnType,
-} from './actions/test/getTxpoolStatus.js'
-export {
-  impersonateAccount,
-  type ImpersonateAccountParameters,
-} from './actions/test/impersonateAccount.js'
-export {
-  increaseTime,
-  type IncreaseTimeParameters,
-} from './actions/test/increaseTime.js'
-export {
-  inspectTxpool,
-  type InspectTxpoolReturnType,
-} from './actions/test/inspectTxpool.js'
-export { mine, type MineParameters } from './actions/test/mine.js'
-export { removeBlockTimestampInterval } from './actions/test/removeBlockTimestampInterval.js'
-export { reset, type ResetParameters } from './actions/test/reset.js'
-export { revert, type RevertParameters } from './actions/test/revert.js'
-export {
-  sendUnsignedTransaction,
-  type SendUnsignedTransactionParameters,
-  type SendUnsignedTransactionReturnType,
-} from './actions/test/sendUnsignedTransaction.js'
-export { setAutomine } from './actions/test/setAutomine.js'
-export {
-  setBalance,
-  type SetBalanceParameters,
-} from './actions/test/setBalance.js'
-export {
-  setBlockGasLimit,
-  type SetBlockGasLimitParameters,
-} from './actions/test/setBlockGasLimit.js'
-export {
-  setBlockTimestampInterval,
-  type SetBlockTimestampIntervalParameters,
-} from './actions/test/setBlockTimestampInterval.js'
-export { setCode, type SetCodeParameters } from './actions/test/setCode.js'
-export {
-  setCoinbase,
-  type SetCoinbaseParameters,
-} from './actions/test/setCoinbase.js'
-export {
-  setIntervalMining,
-  type SetIntervalMiningParameters,
-} from './actions/test/setIntervalMining.js'
-export { setLoggingEnabled } from './actions/test/setLoggingEnabled.js'
-export {
-  setMinGasPrice,
-  type SetMinGasPriceParameters,
-} from './actions/test/setMinGasPrice.js'
-export {
-  setNextBlockBaseFeePerGas,
-  type SetNextBlockBaseFeePerGasParameters,
-} from './actions/test/setNextBlockBaseFeePerGas.js'
-export {
-  setNextBlockTimestamp,
-  type SetNextBlockTimestampParameters,
-} from './actions/test/setNextBlockTimestamp.js'
-export { setNonce, type SetNonceParameters } from './actions/test/setNonce.js'
-export {
-  setStorageAt,
-  type SetStorageAtParameters,
-} from './actions/test/setStorageAt.js'
-export { snapshot } from './actions/test/snapshot.js'
-export { setRpcUrl } from './actions/test/setRpcUrl.js'
-export {
-  stopImpersonatingAccount,
-  type StopImpersonatingAccountParameters,
-} from './actions/test/stopImpersonatingAccount.js'
