+++ src/test/package.json
@@ -1,0 +1,6 @@
+{
+  "type": "module",
+  "types": "../_types/test/index.d.ts",
+  "module": "../_esm/test/index.js",
+  "main": "../_cjs/test/index.js"
+}
