+++ src/wallet/package.json
@@ -1,0 +1,6 @@
+{
+  "type": "module",
+  "types": "../_types/wallet/index.d.ts",
+  "module": "../_esm/wallet/index.js",
+  "main": "../_cjs/wallet/index.js"
+}
