--- src/window.ts
@@ -1,1 +1,0 @@
-import './types/window.js'
