+++ src/window/window.ts
@@ -1,0 +1,1 @@
+import '../types/window.js'
